+++ lib/fastly/models/billing_estimate_lines_line_items.rb
@@ -12,41 +12,37 @@ require 'date'
 require 'time'
 
 module Fastly
-  class BillingResponse
-    # Date and time in ISO 8601 format.
-    attr_accessor :end_time
+  class BillingEstimateLinesLineItems
+    attr_accessor :plan_no
 
-    # Date and time in ISO 8601 format.
-    attr_accessor :start_time
+    attr_accessor :description
 
-    attr_accessor :invoice_id
+    attr_accessor :units
 
-    attr_accessor :customer_id
+    attr_accessor :per_unit_cost
 
-    # The current state of our third-party billing vendor. One of `up` or `down`.
-    attr_accessor :vendor_state
+    attr_accessor :service_no
 
-    attr_accessor :status
+    attr_accessor :service_type
 
-    attr_accessor :total
+    attr_accessor :amount
 
-    # Breakdown of regional data for products that are region based.
-    attr_accessor :regions
+    attr_accessor :client_service_id
 
-    attr_accessor :line_items
+    attr_accessor :client_plan_id
 
     # Attribute mapping from ruby-style variable name to JSON key.
     def self.attribute_map
       {
-        :'end_time' => :'end_time',
-        :'start_time' => :'start_time',
-        :'invoice_id' => :'invoice_id',
-        :'customer_id' => :'customer_id',
-        :'vendor_state' => :'vendor_state',
-        :'status' => :'status',
-        :'total' => :'total',
-        :'regions' => :'regions',
-        :'line_items' => :'line_items'
+        :'plan_no' => :'plan_no',
+        :'description' => :'description',
+        :'units' => :'units',
+        :'per_unit_cost' => :'per_unit_cost',
+        :'service_no' => :'service_no',
+        :'service_type' => :'service_type',
+        :'amount' => :'amount',
+        :'client_service_id' => :'client_service_id',
+        :'client_plan_id' => :'client_plan_id'
       }
     end
 
@@ -58,87 +54,73 @@ module Fastly
     # Attribute type mapping.
     def self.fastly_types
       {
-        :'end_time' => :'Time',
-        :'start_time' => :'Time',
-        :'invoice_id' => :'String',
-        :'customer_id' => :'String',
-        :'vendor_state' => :'String',
-        :'status' => :'BillingStatus',
-        :'total' => :'BillingTotal',
-        :'regions' => :'Hash<String, Hash<String, Object>>',
-        :'line_items' => :'Array<BillingResponseLineItem>'
+        :'plan_no' => :'Integer',
+        :'description' => :'String',
+        :'units' => :'Float',
+        :'per_unit_cost' => :'Float',
+        :'service_no' => :'Float',
+        :'service_type' => :'String',
+        :'amount' => :'Float',
+        :'client_service_id' => :'String',
+        :'client_plan_id' => :'String'
       }
     end
 
     # List of attributes with nullable: true
     def self.fastly_nullable
       Set.new([
-        :'end_time',
-        :'start_time',
       ])
-    end
-
-    # List of class defined in allOf (OpenAPI v3)
-    def self.fastly_all_of
-      [
-      :'Billing',
-      :'BillingResponseItemItemsData'
-      ]
     end
 
     # Initializes the object
     # @param [Hash] attributes Model attributes in the form of hash
     def initialize(attributes = {})
       if (!attributes.is_a?(Hash))
-        fail ArgumentError, "The input argument (attributes) must be a hash in `Fastly::BillingResponse` initialize method"
+        fail ArgumentError, "The input argument (attributes) must be a hash in `Fastly::BillingEstimateLinesLineItems` initialize method"
       end
 
       # check to see if the attribute exists and convert string to symbol for hash key
       attributes = attributes.each_with_object({}) { |(k, v), h|
         if (!self.class.attribute_map.key?(k.to_sym))
-          fail ArgumentError, "`#{k}` is not a valid attribute in `Fastly::BillingResponse`. Please check the name to make sure it's valid. List of attributes: " + self.class.attribute_map.keys.inspect
+          fail ArgumentError, "`#{k}` is not a valid attribute in `Fastly::BillingEstimateLinesLineItems`. Please check the name to make sure it's valid. List of attributes: " + self.class.attribute_map.keys.inspect
         end
         h[k.to_sym] = v
       }
 
-      if attributes.key?(:'end_time')
-        self.end_time = attributes[:'end_time']
+      if attributes.key?(:'plan_no')
+        self.plan_no = attributes[:'plan_no']
       end
 
-      if attributes.key?(:'start_time')
-        self.start_time = attributes[:'start_time']
+      if attributes.key?(:'description')
+        self.description = attributes[:'description']
       end
 
-      if attributes.key?(:'invoice_id')
-        self.invoice_id = attributes[:'invoice_id']
+      if attributes.key?(:'units')
+        self.units = attributes[:'units']
       end
 
-      if attributes.key?(:'customer_id')
-        self.customer_id = attributes[:'customer_id']
+      if attributes.key?(:'per_unit_cost')
+        self.per_unit_cost = attributes[:'per_unit_cost']
       end
 
-      if attributes.key?(:'vendor_state')
-        self.vendor_state = attributes[:'vendor_state']
+      if attributes.key?(:'service_no')
+        self.service_no = attributes[:'service_no']
       end
 
-      if attributes.key?(:'status')
-        self.status = attributes[:'status']
+      if attributes.key?(:'service_type')
+        self.service_type = attributes[:'service_type']
       end
 
-      if attributes.key?(:'total')
-        self.total = attributes[:'total']
+      if attributes.key?(:'amount')
+        self.amount = attributes[:'amount']
       end
 
-      if attributes.key?(:'regions')
-        if (value = attributes[:'regions']).is_a?(Hash)
-          self.regions = value
-        end
+      if attributes.key?(:'client_service_id')
+        self.client_service_id = attributes[:'client_service_id']
       end
 
-      if attributes.key?(:'line_items')
-        if (value = attributes[:'line_items']).is_a?(Array)
-          self.line_items = value
-        end
+      if attributes.key?(:'client_plan_id')
+        self.client_plan_id = attributes[:'client_plan_id']
       end
     end
 
@@ -160,15 +142,15 @@ module Fastly
     def ==(o)
       return true if self.equal?(o)
       self.class == o.class &&
-          end_time == o.end_time &&
-          start_time == o.start_time &&
-          invoice_id == o.invoice_id &&
-          customer_id == o.customer_id &&
-          vendor_state == o.vendor_state &&
-          status == o.status &&
-          total == o.total &&
-          regions == o.regions &&
-          line_items == o.line_items
+          plan_no == o.plan_no &&
+          description == o.description &&
+          units == o.units &&
+          per_unit_cost == o.per_unit_cost &&
+          service_no == o.service_no &&
+          service_type == o.service_type &&
+          amount == o.amount &&
+          client_service_id == o.client_service_id &&
+          client_plan_id == o.client_plan_id
     end
 
     # @see the `==` method
@@ -180,7 +162,7 @@ module Fastly
     # Calculates hash code according to all attributes.
     # @return [Integer] Hash code
     def hash
-      [end_time, start_time, invoice_id, customer_id, vendor_state, status, total, regions, line_items].hash
+      [plan_no, description, units, per_unit_cost, service_no, service_type, amount, client_service_id, client_plan_id].hash
     end
 
     # Builds the object from hash