+++ lib/fastly/models/billing_response_item_items_data.rb
@@ -12,20 +12,13 @@ require 'date'
 require 'time'
 
 module Fastly
-  class UserResponseAllOf
-    attr_accessor :id
-
-    # The alphanumeric string identifying a email login.
-    attr_accessor :email_hash
-
-    attr_accessor :customer_id
+  class BillingResponseItemItemsData
+    attr_accessor :line_items
 
     # Attribute mapping from ruby-style variable name to JSON key.
     def self.attribute_map
       {
-        :'id' => :'id',
-        :'email_hash' => :'email_hash',
-        :'customer_id' => :'customer_id'
+        :'line_items' => :'line_items'
       }
     end
 
@@ -37,9 +30,7 @@ module Fastly
     # Attribute type mapping.
     def self.fastly_types
       {
-        :'id' => :'String',
-        :'email_hash' => :'String',
-        :'customer_id' => :'String'
+        :'line_items' => :'Array<BillingResponseLineItem>'
       }
     end
 
@@ -53,27 +44,21 @@ module Fastly
     # @param [Hash] attributes Model attributes in the form of hash
     def initialize(attributes = {})
       if (!attributes.is_a?(Hash))
-        fail ArgumentError, "The input argument (attributes) must be a hash in `Fastly::UserResponseAllOf` initialize method"
+        fail ArgumentError, "The input argument (attributes) must be a hash in `Fastly::BillingResponseItemItemsData` initialize method"
       end
 
       # check to see if the attribute exists and convert string to symbol for hash key
       attributes = attributes.each_with_object({}) { |(k, v), h|
         if (!self.class.attribute_map.key?(k.to_sym))
-          fail ArgumentError, "`#{k}` is not a valid attribute in `Fastly::UserResponseAllOf`. Please check the name to make sure it's valid. List of attributes: " + self.class.attribute_map.keys.inspect
+          fail ArgumentError, "`#{k}` is not a valid attribute in `Fastly::BillingResponseItemItemsData`. Please check the name to make sure it's valid. List of attributes: " + self.class.attribute_map.keys.inspect
         end
         h[k.to_sym] = v
       }
 
-      if attributes.key?(:'id')
-        self.id = attributes[:'id']
-      end
-
-      if attributes.key?(:'email_hash')
-        self.email_hash = attributes[:'email_hash']
-      end
-
-      if attributes.key?(:'customer_id')
-        self.customer_id = attributes[:'customer_id']
+      if attributes.key?(:'line_items')
+        if (value = attributes[:'line_items']).is_a?(Array)
+          self.line_items = value
+        end
       end
     end
 
@@ -95,9 +80,7 @@ module Fastly
     def ==(o)
       return true if self.equal?(o)
       self.class == o.class &&
-          id == o.id &&
-          email_hash == o.email_hash &&
-          customer_id == o.customer_id
+          line_items == o.line_items
     end
 
     # @see the `==` method
@@ -109,7 +92,7 @@ module Fastly
     # Calculates hash code according to all attributes.
     # @return [Integer] Hash code
     def hash
-      [id, email_hash, customer_id].hash
+      [line_items].hash
     end
 
     # Builds the object from hash