+++ lib/fastly/models/user_response.rb
@@ -99,6 +99,9 @@ module Fastly
     # List of attributes with nullable: true
     def self.fastly_nullable
       Set.new([
+        :'locked',
+        :'require_new_password',
+        :'two_factor_auth_enabled',
         :'created_at',
         :'deleted_at',
         :'updated_at',
@@ -110,7 +113,7 @@ module Fastly
       [
       :'Timestamps',
       :'User',
-      :'UserResponseAllOf'
+      :'UserResponseReadOnly'
       ]
     end